♻️ vue store 재구성 (#43)
- actions 추가

Related to: #43
