+++ vue/src/store/getters.js
@@ -1,7 +1,4 @@
 export default {
     getUserName: state => state.userName,
-    getToken: state => state.token,
-    isLogin(state) {
-        return state.token != null;
-    }
+    getToken: state => state.token
 }
