+++ vue/src/store/store.js
@@ -1,6 +1,7 @@
 import {createStore} from "vuex";
 import mutations from "@/store/mutations";
 import getters from "@/store/getters";
+import actions from "@/store/actions";
 
 const store = createStore({
     state: { // 변수 집합
@@ -9,6 +10,7 @@ const store = createStore({
     },
     getters, // 변수 get
     mutations, // 변수 set
+    actions,
 });
 
 export default store;
